commons-rdf-simple module
dummyimpl -> simple
