+++ commons-rdf-simple/src/main/java/com/github/commonsrdf/simple/DummyRDFTermFactoryTest.java
@@ -11,16 +11,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.github.commonsrdf.dummyimpl;
+package com.github.commonsrdf.simple;
 
-import com.github.commonsrdf.api.AbstractGraphTest;
+import com.github.commonsrdf.api.AbstractRDFTermFactoryTest;
 import com.github.commonsrdf.api.RDFTermFactory;
 
-/**
- * Test DummyRDFTermFactory with AbstractGraphTest
- *
- */
-public class DummyGraphTest extends AbstractGraphTest {
+public class DummyRDFTermFactoryTest extends AbstractRDFTermFactoryTest {
 
 	@Override
 	public RDFTermFactory createFactory() {
